Move grpc tests
